+++ eventkviz.php
@@ -16,7 +16,7 @@
  * Plugin Name:       Eventkviz
  * Plugin URI:        http://eventkviz.sk/
  * Description:       Quizes for events
- * Version:           1.2.4
+ * Version:           1.3.0
  * Author:            Maros Markovic
  * Author URI:        http://eventkviz.sk/
  * License:           GPL-2.0+
@@ -35,7 +35,7 @@ if ( ! defined( 'WPINC' ) ) {
  * Start at version 1.0.0 and use SemVer - https://semver.org
  * Rename this for your plugin and update it as you release new versions.
  */
-define( 'EVENKVIZ_VERSION', '1.2.4' );
+define( 'EVENKVIZ_VERSION', '1.3.0' );
 
 /**
  * The code that runs during plugin activation.
@@ -74,6 +74,10 @@ require_once( plugin_dir_path( __FILE__ ) . 'includes/class-eventkviz-statistika
 require_once( plugin_dir_path( __FILE__ ) . 'includes/class-eventkviz-seedpage.php' );
 require_once( plugin_dir_path( __FILE__ ) . 'includes/class-eventkviz-finalpage.php' );
 require_once( plugin_dir_path( __FILE__ ) . 'includes/class-eventkviz-rest.php' );
+if ( is_admin() ) {
+	require_once( plugin_dir_path( __FILE__ ) . 'admin/class-eventkviz-leaderboard.php' );
+	Eventkviz_Leaderboard::init();
+}
 
 Eventkviz_Rest_Search::init();